requirements.txt
python-gnupg==0.5.3
tabulate==0.9.0
